telegram_assistant/requirements.txt
python-telegram-bot[job-queue]==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
APScheduler==3.10.4
SQLAlchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2
fastapi==0.110.0
uvicorn==0.29.0
